requirements.txt
pandas
scikit-learn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
beautifulsoup4
PyYAML